refactor: remove old/depreciated AppConfig folder
