+++ src/components/settings/integrations/index.tsx
@@ -2,17 +2,17 @@ import { CheckOutlined, ExclamationOutlined } from '@ant-design/icons'
 import { Avatar, Button, List, Tag } from 'antd'
 // import update from 'immutability-helper'
 import { ReactNode, useState } from 'react'
-import ExchangeRateLogo from '../../assets/integrationsKeysIcon/ExchangeRateService.svg?react'
-import SegmentLogo from '../../assets/integrationsKeysIcon/Segment.svg?react'
-import SendGridLogo from '../../assets/integrationsKeysIcon/SendGrid.svg?react'
-import UniBeeLogo from '../../assets/integrationsKeysIcon/UniBeeKeys.svg?react'
-import VATsenseLogo from '../../assets/integrationsKeysIcon/VATsense.svg?react'
-import { useAppConfigStore } from '../../stores'
-import UniBeeAPIKeyModal from './appConfig/apiKeyModal'
-import ExchangeRateModal from './appConfig/exchangeRateKeyModal'
-import SegmentModal from './appConfig/segmentModal'
-import SendGridModal from './appConfig/sendGridKeyModal'
-import VATModal from './appConfig/vatKeyModal'
+import ExchangeRateLogo from '../../../assets/integrationsKeysIcon/ExchangeRateService.svg?react'
+import SegmentLogo from '../../../assets/integrationsKeysIcon/Segment.svg?react'
+import SendGridLogo from '../../../assets/integrationsKeysIcon/SendGrid.svg?react'
+import UniBeeLogo from '../../../assets/integrationsKeysIcon/UniBeeKeys.svg?react'
+import VATsenseLogo from '../../../assets/integrationsKeysIcon/VATsense.svg?react'
+import { useAppConfigStore } from '../../../stores'
+import UniBeeAPIKeyModal from './UniBeeAPIKeyModal'
+import ExchangeRateModal from './exchangeRateKeyModal'
+import SegmentModal from './segmentModal'
+import SendGridModal from './sendGridKeyModal'
+import VATModal from './vatKeyModal'
 
 type TAPP_Integration = {
   IsSetupFinished: boolean
